Fix broken links on package home pages
This is achieved by introducing a PageStatus context component that
notifies the SwitchLink component when it is inside a page that has been
moved from a README file into a directory URL without the readme path.
The SwitchLink can then factor that into the href manipulation to ensure
relative links still work correctly.