Use common getTitle method to print page titles
Even though we always generate a default meta.title entry for DocPages,
other types of pages don't have an entry. Therefore we are using a
common getTitle function to ensure we are always retrieving the title
from the correct place and that it is formatted correctly.
This means removing some formatting from tree nav content which would
erroneously titleCase custom titles.